backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.10.3
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
reportlab==4.2.5
